Combined the functions of dynamics and statics into one

+++ mihalchenko.vladislav/P4/inputArray.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 #include "inputArray.hpp"
 
-size_t mihalchenko::inputArray(std::istream& in, int* arrInput, size_t arrCols)
+size_t mihalchenko::inputArray(std::istream& in, float* arrInput, size_t arrCols)
 {
   for (size_t i = 0; i < arrCols; ++i)
   {
